Add ticket compensation feature and update task creation parameters
- Introduced parser_compensation_set and parser_compensation_get functions to manage ticket compensation settings.
- Updated task_create to use float for yield_fraction instead of double for consistency.
- Modified task_apply_compensation to handle float yield_fraction.
- Enhanced scheduler_main_loop to apply compensation based on task yield status.
- Refactored task_do_pi_calc functions to utilize early limit calculation based on yield_fraction.

+++ src/parser.c
@@ -204,7 +204,7 @@ int parser_load(const char *filename, struct node **head)
          * NOTE: This should be freed inside the
          * double linked list infraestructure.
          */
-        task_t *task = task_create(id, tickets, work_units, yield_fraction);
+        task_t *task = task_create(id, tickets, work_units, (float)yield_fraction);
         if(!dll_insert_node(head, task, id, tickets, work_units)) {
             fprintf(stderr, "Parser: failed to insert node at line %d: '%s'\n", lineno, line);
             goto err;
